+++ parser_library/src/workspaces/workspace.h
@@ -124,7 +124,7 @@ public:
 
     bool settings_updated();
 
-    const processor_group& get_proc_grp_by_program(const resource_location& file) const;
+    const processor_group& get_proc_grp(const resource_location& file) const;
     const processor_group& get_proc_grp(const proc_grp_id& id) const; // test only
 
     std::vector<std::pair<std::string, size_t>> make_opcode_suggestion(